% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_doj_race_cats.R
\name{is_census_key_set}
\alias{is_census_key_set}
\title{Check if a Census API Key is set}
\usage{
is_census_key_set()
}
\value{
Logical TRUE if set, FALSE if not
}
\description{
Check if a Census API Key is set
}
